bugfix(extract): only add cycles in summary if the rule contains a circular attribute
