+++ src/utl/findRuleByName.js
@@ -5,6 +5,7 @@ const _get = require("lodash/get");
  * and undefined if no such rule exists/ the rule is an 'allowed'
  * rule.
  *
+ * (this thing probably belongs in a model-like folder and not in utl)
  * @param {IRuleSetType} pRuleSet - The rule set to search in
  * @param {string} pName - The rule name to look for
  * @return {IForbiddenRuleType|IAllowedRuleType} - a rule (or 'undefined' if nothing found)
